+++ pkcs11/src/backend/object.rs
@@ -1,12 +1,12 @@
 use cryptoki_sys::{CKA_CLASS, CKA_ID, CKA_LABEL, CK_OBJECT_CLASS, CK_SESSION_HANDLE};
-use log::{debug, trace, warn};
+use log::{debug, trace};
 
 use super::{
     db::{
         attr::{CkRawAttr, CkRawAttrTemplate},
         object::ObjectKind,
     },
-    key::Id,
+    key::{NetHSMId, Pkcs11Id},
     session::Session,
     Error,
 };
@@ -21,8 +21,7 @@ pub struct EnumCtx {
 #[derive(Clone, Debug)]
 pub struct KeyRequirements {
     pub kind: Option<ObjectKind>,
-    pub id: Option<String>,
-    pub invalid_id: bool,
+    pub id: Option<NetHSMId>,
 }
 
 fn parse_key_requirements(template: Option<CkRawAttrTemplate>) -> Result<KeyRequirements, Error> {
@@ -30,47 +29,34 @@ fn parse_key_requirements(template: Option<CkRawAttrTemplate>) -> Result<KeyRequ
         Some(template) => {
             let mut key_id = None;
             let mut kind = None;
-            let mut invalid_id = false;
             for attr in template.iter() {
                 debug!("attr {:?}: {:?}", attr.type_(), attr.val_bytes());
 
                 if attr.type_() == CKA_CLASS {
                     kind = unsafe { attr.read_value::<CK_OBJECT_CLASS>() }.map(ObjectKind::from)
                 }
-
                 if attr.type_() == CKA_ID {
-                    if let Some(bytes) = attr.val_bytes() {
-                        if let Ok(id) = Id::try_from(bytes.to_owned()) {
-                            key_id = Some(id.into());
-                        } else {
-                            warn!("Invalid ID in key requirements");
-                            invalid_id = true;
-                        }
-                    }
+                    key_id = Some(parse_nethsm_id(&attr)?);
                 }
                 if attr.type_() == CKA_LABEL && key_id.is_none() {
-                    key_id = Some(parse_str_from_attr(&attr)?);
+                    key_id = Some(parse_nethsm_id(&attr)?);
                 }
             }
-            Ok(KeyRequirements {
-                kind,
-                id: key_id,
-                invalid_id,
-            })
+            Ok(KeyRequirements { kind, id: key_id })
         }
         None => Ok(KeyRequirements {
             kind: None,
             id: None,
-            invalid_id: false,
         }),
     }
 }
 
-fn parse_str_from_attr(attr: &CkRawAttr) -> Result<String, Error> {
+fn parse_nethsm_id(attr: &CkRawAttr) -> Result<NetHSMId, Error> {
     let bytes = attr
         .val_bytes()
         .ok_or(Error::InvalidAttribute(attr.type_()))?;
-    Ok(String::from_utf8(bytes.to_vec())?)
+    let pkcs11_id = Pkcs11Id::from(bytes);
+    NetHSMId::try_from(&pkcs11_id).map_err(|_| Error::InvalidAttribute(attr.type_()))
 }
 
 impl EnumCtx {
@@ -79,13 +65,7 @@ impl EnumCtx {
         template: Option<CkRawAttrTemplate>,
     ) -> Result<Self, Error> {
         let key_req = parse_key_requirements(template)?;
-
-        // Elements with a non-compliant ID are no longer supported
-        let handles = if key_req.invalid_id {
-            vec![]
-        } else {
-            session.find_key(key_req.id.as_deref(), key_req.kind)?
-        };
+        let handles = session.find_key(key_req.id.as_ref(), key_req.kind)?;
         Ok(EnumCtx::new(handles))
     }
 
@@ -121,7 +101,6 @@ mod tests {
 
         assert_eq!(res.kind, None);
         assert_eq!(res.id, None);
-        assert!(!res.invalid_id);
 
         Ok(())
     }
@@ -144,8 +123,7 @@ mod tests {
         let res = parse_key_requirements(template)?;
 
         assert_eq!(res.kind, None);
-        assert_eq!(res.id, None);
-        assert!(res.invalid_id);
+        assert_eq!(res.id.as_ref().map(|id| id.as_str()), Some("0---00FF00FF"));
 
         Ok(())
     }
@@ -168,8 +146,7 @@ mod tests {
         let res = parse_key_requirements(template)?;
 
         assert_eq!(res.kind, None);
-        assert_eq!(res.id, Some("test".to_string()));
-        assert!(!res.invalid_id);
+        assert_eq!(res.id.as_ref().map(|id| id.as_str()), Some("test"));
 
         Ok(())
     }